backend/requirements.txt
Flask==1.1.2
numpy==1.19.5
Pillow==8.4.0
tensorflow==2.6.2
gunicorn==20.1.0